Major version update to 1.5.0. Move additions, ability additions + changes, addition of Regigigas, addition of Castform Sandstorm form, addition of HM items with 'boss' wild pokemon battles to get them, addition of custom 'Weather Orb' item that randomly changes weather every 3 turns, ability of all pokemon to be obtained. Battle Frontier allows for 1 restricted pokemon on teams, and many, many more changes.

+++ src/easy_chat.c
@@ -4716,7 +4716,7 @@ static void sub_811E7F8(void)
 
 static void sub_811E828(void)
 {
-    int graphicsId;
+    u16 graphicsId;
     u8 spriteId;
 
     switch (GetDisplayedPersonType())
